Reject out-of-range integers in the Ruby and PHP extensions (#6069)
(cherry picked from commit 08ae7fd8d5d60f46979f8a4421f25237767a1e53)

+++ ruby/src/IceRuby/Operation.cpp
@@ -368,7 +368,7 @@ IceRuby::OperationI::convertParam(VALUE v, long pos)
     ParamInfoPtr param = make_shared<ParamInfo>();
     param->type = getType(RARRAY_AREF(v, 0));
     param->optional = static_cast<bool>(RTEST(RARRAY_AREF(v, 1)));
-    param->tag = static_cast<int>(getInteger(RARRAY_AREF(v, 2)));
+    param->tag = getInteger(RARRAY_AREF(v, 2));
     param->pos = static_cast<int>(pos);
     return param;
 }
